moved some components into their own folders, socket widgets broken
