+++ client/components/navbar/index.js
@@ -1,0 +1,3 @@
+import Navbar from './navbar'
+
+export default Navbar
