+++ client/components/navbar/navbar.spec.js
