+++ client/components/twitch-login/index.js
@@ -1,0 +1,3 @@
+import TwitchLogin from './twitch-login'
+
+export default TwitchLogin
