+++ client/components/welcome-text/index.js
@@ -1,0 +1,2 @@
+import WelcomeText from './welcome-text'
+export default (WelcomeText)
