+++ client/components/welcome-text/welcome-text.js
