+++ client/components/welcome-text/welcome-text.spec.js
