+++ client/components/widgets/gameboy/index.js
@@ -1,0 +1,4 @@
+import GameboyDashboard, {gameboyEvents} from './gameboy-dashboard'
+import GameboyOverlay from './gameboy-overlay'
+
+export {GameboyDashboard, gameboyEvents, GameboyOverlay}
