+++ client/components/widgets/whiteboard/index.js
@@ -1,0 +1,4 @@
+import WhiteboardDashboard, {whiteboardEvents} from './whiteboard-dashboard'
+import WhiteboardOverlay from './whiteboard-overlay'
+
+export {WhiteboardDashboard, whiteboardEvents, WhiteboardOverlay}
